requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
amazon-bedrock-haystack==1.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
asn1crypto==1.5.1
astra-haystack==0.10.0
astrapy==1.5.2
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.7
attrs==24.2.0
Authlib==1.3.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
boilerpy3==1.0.7
boto3==1.35.68
botocore==1.35.68
build==1.2.2.post1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
chainlit==2.1.1
chardet==5.2.0
charset-normalizer==3.4.0
chevron==0.14.0
chroma-haystack==1.0.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
compressed-tensors==0.6.0
cryptography==43.0.3
dataclasses-json==0.6.7
datasets==3.1.0
decorator==5.1.1
deepeval==0.20.57
deepeval-haystack==0.1.2
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
durationpy==0.9
einops==0.8.0
elastic-transport==8.15.1
elasticsearch==8.16.0
elasticsearch-haystack==1.0.1
emoji==2.14.0
eval_type_backport==0.2.0
Events==0.5
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
fastapi==0.112.4
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
google-auth==2.36.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
haystack-ai==2.9.0
haystack-experimental==0.3.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
iniconfig==2.0.0
interegular==0.3.3
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.21
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.139
lark==1.2.2
Lazify==0.4.0
lazy-imports==0.3.1
literalai==0.0.623
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.4.4
mmh3==5.0.1
mongodb-atlas-haystack==1.0.0
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-ml-py==12.560.30
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.54.1
opencv-python-headless==4.10.0.84
opensearch-haystack==1.1.0
opensearch-py==2.7.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-exporter-otlp-proto-http==1.28.0
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.11
outlines==0.0.46
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pgvector==0.3.6
pgvector-haystack==1.2.0
pillow==10.4.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone_haystack==2.0.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
prompthub-py==4.0.0
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyOpenSSL==24.2.1
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
pytest==8.3.3
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.1
python-socketio==5.11.4
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.12.1
qdrant-haystack==7.0.0
quantulum3==0.9.2
ragas==0.1.16
ragas-haystack==1.0.1
rank-bm25==0.2.2
RapidFuzz==3.10.1
ray==2.38.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==0.9.8
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.19.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.3
snowflake-haystack==0.0.2
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.3
syncer==2.0.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.3
typeguard==4.4.1
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.6
unstructured-client==0.28.0
unstructured-fileconverter-haystack==0.4.1
uptrace==1.27.0
url-normalize==1.4.3
urllib3==2.2.3
uuid6==2024.7.10
uvicorn==0.25.0
validators==0.34.0
watchfiles==0.20.0
wcwidth==0.2.13
weaviate-client==4.9.4
weaviate-haystack==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.20.2
